Moved EventSourceFactory and ServiceMonitorClient to -core
