+++ shard-operator-parent/shard-operator-core/pom.xml
@@ -40,6 +40,10 @@
       <artifactId>infra-v1</artifactId>
     </dependency>
     <dependency>
+      <groupId>io.quarkiverse.operatorsdk</groupId>
+      <artifactId>quarkus-operator-sdk</artifactId>
+    </dependency>
+    <dependency>
       <groupId>io.quarkiverse.loggingjson</groupId>
       <artifactId>quarkus-logging-json</artifactId>
     </dependency>
